--- client/src/assets/css_group/App.module.css
@@ -1,203 +1,0 @@
-:local .title {
-  grid-column-start: 2;
-  grid-row-start: 2;
-
-  justify-self: center;
-}
-
-:local .startBtn {
-  grid-column-start: 2;
-  grid-row-start: 2;
-  width: 50px;
-  height: 50px;
-  background-color: #61dafb;
-  border-radius: 15px;
-  border: 5px solid rgb(144, 210, 245);
-  transition: 0.25s ease-in-out;
-  animation: App-logo-spin infinite 20s linear;
-  min-width: 50px;
-  align-self: center;
-  justify-self: center;
-}
-
-:local .startBtn:hover {
-  background-color: #00bdf1;
-  border: 5px solid rgb(94, 136, 242);
-}
-
-:local .startBtnActivated {
-  background-color: rgb(173, 255, 169);
-  border: 5px solid black;
-  opacity: 1;
-}
-
-:local .startBtnBuffer {
-  opacity: 0.6;
-}
-
-:local .App {
-  height: 100vh;
-  width: 100vw;
-  display: grid;
-  grid-template-rows: 0.1fr 1fr 1fr;
-  grid-template-columns: 1fr 1fr 1fr;
-  flex-direction: column;
-}
-
-:local .engineer {
-  margin-top: 50px;
-  grid-row: 2;
-  align-self: flex-end;
-  grid-column: 2;
-}
-
-/* LoggedIn page - start */
-
-:local .loggedIn {
-  width: 100vw;
-  height: 100vh;
-  display: grid;
-  grid-template-rows: 0.05fr 1fr;
-}
-
-:global .loggedInNavbar {
-  padding: 0;
-  margin: 0;
-  grid-row-start: 1;
-  display: flex;
-  gap: 8px;
-  justify-content: flex-end;
-  align-items: flex-end;
-}
-
-/* LoggedIn page  - End */
-
-/* Logout button - start */
-
-:local .logoutBtn {
-  border: 1px solid rgb(0, 0, 0);
-  width: 50px;
-  margin: 0;
-  height: 50px;
-  border-radius: 25px;
-  background-image: url("https://cdn.discordapp.com/attachments/463174949310562304/1075952796660146196/IMG_3462.png");
-  background-size: cover;
-  min-width: 50px;
-  transition: 0.25s ease-in-out;
-  font-family: "PP Neue Montreal";
-}
-
-:local .logoutBtnClicked {
-  border: 1px solid #00bdf1;
-}
-
-:local .logoutBtn:hover,
-:local .logoutBtnClicked:hover {
-  opacity: 0.8;
-}
-
-:local .logoutPop {
-  position: absolute;
-  right: 10px;
-  margin-top: 10px;
-  grid-template-rows: 1fr 1fr 1fr;
-  box-shadow: 0px 0px 5px rgba(0, 0, 0, 0.363);
-  background-color: white;
-  border-radius: 25px;
-  display: flex;
-  flex-direction: column;
-}
-
-:local .logoutPop button p {
-  font-family: "Roboto", sans-serif;
-  font-weight: 500;
-  color: black;
-}
-
-:local .logoutPop button:hover {
-  background-color: rgba(8, 8, 8, 0.096);
-}
-
-:local .logoutBtnCover {
-  border-radius: 25px;
-  width: 50px;
-  margin-right: 10px;
-  margin-top: 10px;
-}
-
-:local .individualBtn {
-  background-color: transparent;
-  color: none;
-  border-radius: 25px;
-  display: flex;
-  align-items: center;
-  gap: 20px;
-  margin: 0;
-  transition: 0.25s ease-in-out;
-  cursor: pointer;
-}
-
-:local .individualBtn svg {
-  height: 15px;
-  color: black;
-  width: 15px;
-}
-
-/* Logout button - end */
-
-/* Login button - start */
-
-:local .loginBtn,
-:local .signupBtn {
-  transition: 0.25s ease-in-out;
-  width: 200px;
-  height: 50px;
-  font-family: "PP Neue Montreal";
-  background-color: black;
-}
-
-/* Login button - end */
-
-:local .btnWrap {
-  grid-row-start: 3;
-  grid-column-start: 2;
-  display: grid;
-  padding: 20px;
-  place-items: center;
-  /*outline: dashed 1px red; */
-}
-
-/* Login button - start */
-
-:local .loginBtn {
-  width: 250px;
-  height: 50px;
-}
-
-:local .signupBtn:hover,
-:local .loginBtn:hover {
-  opacity: 0.8;
-  /* signUp button - start */
-}
-
-:local .signupBtn {
-  width: 250px;
-  height: 50px;
-  font: PP Neue Montreal;
-}
-
-:local .dashboard_button_example {
-  width: 250px;
-  height: 50px;
-  font: PP Neue Montreal;
-}
-
-/* signUp button - end */
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
