+++ client/src/assets/css_group/fonts.module.css
@@ -1,7 +1,6 @@
 @import url("https://fonts.googleapis.com/css2?family=Roboto&display=swap");
-:local .myFont {
-  @font-face {
-    font-family: "MyFont";
-    src: url("./fonts/MyFont.ttf") format("truetype");
-  }
+
+@font-face {
+  font-family: "MyFont";
+  src: url("./fonts/MyFont.ttf") format("truetype");
 }
