--- client/src/assets/css_group/form.module.css
@@ -1,138 +1,0 @@
-:local p,
-div {
-  font-family: PP Neue Montreal;
-}
-
-:local .logintitle {
-  font-size: 24px;
-  line-height: 60px;
-  float: left;
-  font-weight: bold;
-  font-family: PP Neue Montreal;
-  text-align: center;
-}
-
-:local body {
-  margin: 0;
-  padding: 0;
-}
-
-:local .linkstyle {
-  text-decoration: none;
-  text-align: center;
-  color: black;
-  font-size: smaller;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 15px;
-}
-
-:local .googleSignIn {
-  width: 100%;
-
-  margin-top: 10px;
-}
-
-:local .wrapper {
-  height: 100vh;
-  width: 100vw;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  background-color: rgb(233, 246, 253);
-  background-image: url("../image_group/blue-pink-better-theme.png");
-  background-position-y: 40%;
-  opacity: 0.9;
-  background-size: cover;
-}
-
-:local .form-wrapper {
-  display: flex;
-  flex-direction: column;
-  width: 280px;
-  max-width: 80%;
-  min-width: 100px;
-  min-height: 400px;
-  padding: 20px 40px;
-  border-radius: 25px;
-  box-shadow: 0px 8px 36px #222;
-  background-color: #fefefe;
-}
-
-:local .form-wrapper > h2 {
-  display: flex;
-  justify-content: center;
-  font-family: "Segoe UI", "Neue Montreal", "Ubuntu", "Roboto", "Open Sans",
-    "Helvetica Neue", sans-serif;
-  font-size: 2em;
-  font-weight: lighter;
-  margin-top: 0.25em;
-  color: #222;
-}
-
-:local form {
-  width: 100%;
-  display: flex;
-  flex-wrap: wrap;
-}
-
-:local label {
-  margin-bottom: 0.5em;
-  color: #444;
-  font-weight: lighter;
-}
-
-:local input {
-  padding: 15px 15px;
-  border-radius: 15px;
-  outline: none;
-  border: 1px solid #d6d1d5;
-}
-
-:local input::placeholder {
-  font-size: 1.2em;
-  font-weight: lighter;
-  color: #bbb;
-}
-
-:local button {
-  min-width: 100%;
-  cursor: pointer;
-  margin-right: 0.25em;
-  margin-top: 0.5em;
-  padding: 0.938em;
-  border: none;
-  border-radius: 15px;
-  color: white;
-}
-:local .buttonNormal {
-  background-color: black;
-}
-
-:local .buttonColorful {
-  background-image: linear-gradient(to right, #64cfff, #ff59a1);
-}
-
-:local .username,
-:local .email,
-:local .password {
-  display: flex;
-  border-radius: 5px 20px 5px;
-  flex-direction: column;
-  margin-bottom: 15px;
-  width: 100%;
-}
-
-:local .submit {
-  width: 100%;
-  border-radius: 500px;
-  display: flex;
-  flex-wrap: wrap;
-}
-
-:local .centerGoogleAuth {
-  display: grid;
-  margin-top: 20px;
-  margin-left: 40px;
-}
